feat: update rule name
